Generalizing GenEra to make actual Alonzo examples.
got rid of txEraUpdate
A trial to get the WPPHash to agree.
lots of problems with the WPPHash in the TxBody. To compute this I need the Redeemers field of the Witnesses.
To compute the Witnesses, I need the hash of the TxBody, But I can't hash the TxBody without the WPPHash.
fixed typos, added d3 d4to exports
Sets collateral and wppHash, possible infinite loop.
Added TwoPhaseInfo, now Plutus scripts are integrated, and we can get their data and redeemers from TwoPhaseInfo in the DataSpace.
Added ScriptSpace and DataSpace, altered getEraTxOut to someties make Plutus scripts.